+++ src/op_thresholding.cpp
@@ -24,12 +24,12 @@ THE SOFTWARE.
 #include <hip/hip_runtime.h>
 
 #include <functional>
-#include "common/strided_data_wrap.hpp"
+
 #include "common/validation_helpers.hpp"
-#include "core/wrappers/image_wrapper.hpp"
-#include "core/wrappers/generic_tensor_wrapper.hpp"
 #include "core/exception.hpp"
 #include "core/status_type.h"
+#include "core/wrappers/generic_tensor_wrapper.hpp"
+#include "core/wrappers/image_wrapper.hpp"
 #include "kernels/device/thresholding_device.hpp"
 #include "kernels/host/thresholding_host.hpp"
 
@@ -40,74 +40,81 @@ Threshold::Threshold(eThresholdType threshType, int32_t maxBatchSize)
 Threshold::~Threshold() {}
 
 template <typename T>
-void dispatch_threshold_dtype(hipStream_t stream, const Tensor &input, const Tensor &output,
-                                    const Tensor &thresh, const Tensor &maxVal, eThresholdType m_threshType, int32_t m_maxBatchSize, eDeviceType device) {
+void dispatch_threshold_dtype(hipStream_t stream, const Tensor &input, const Tensor &output, const Tensor &thresh,
+                              const Tensor &maxVal, eThresholdType m_threshType, int32_t m_maxBatchSize,
+                              eDeviceType device) {
     ImageWrapper<T> inputWrapper(input);
     ImageWrapper<T> outputWrapper(output);
 
     const auto height = input.shape()[input.shape().layout().height_index()];
     const auto width = input.shape()[input.shape().layout().width_index()];
-    
+
     if (device == eDeviceType::GPU) {
         dim3 block(64, 16);
         dim3 grid((width + block.x - 1) / block.x, (height + block.y - 1) / block.y, m_maxBatchSize);
-        
+
         switch (m_threshType) {
             case THRESH_BINARY:
-                Kernels::Device::binary_generic<<<grid, block, 0, stream>>>(inputWrapper, outputWrapper, 
-                                    GenericTensorWrapper<double>(thresh), GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
+                Kernels::Device::binary_generic<<<grid, block, 0, stream>>>(
+                    inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                    GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
                 break;
             case THRESH_BINARY_INV:
-                Kernels::Device::binary_inv_generic<<<grid, block, 0, stream>>>(inputWrapper, outputWrapper, 
-                                    GenericTensorWrapper<double>(thresh), GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
+                Kernels::Device::binary_inv_generic<<<grid, block, 0, stream>>>(
+                    inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                    GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
                 break;
             case THRESH_TRUNC:
-                Kernels::Device::trunc_generic<<<grid, block, 0, stream>>>(inputWrapper, outputWrapper, 
-                                    GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Device::trunc_generic<<<grid, block, 0, stream>>>(
+                    inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
                 break;
             case THRESH_TOZERO:
-                Kernels::Device::tozero_generic<<<grid, block, 0, stream>>>(inputWrapper, outputWrapper, 
-                                    GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Device::tozero_generic<<<grid, block, 0, stream>>>(
+                    inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
                 break;
             case THRESH_TOZERO_INV:
-                Kernels::Device::tozeroinv_generic<<<grid, block, 0, stream>>>(inputWrapper, outputWrapper, 
-                                    GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Device::tozeroinv_generic<<<grid, block, 0, stream>>>(
+                    inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
                 break;
         }
-        
+
     } else if (device == eDeviceType::CPU) {
-        
         switch (m_threshType) {
             case THRESH_BINARY:
-                Kernels::Host::binary_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), 
-                                GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
+                Kernels::Host::binary_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                                              GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
                 break;
             case THRESH_BINARY_INV:
-                Kernels::Host::binary_inv_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), 
-                                GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
-                break;       
+                Kernels::Host::binary_inv_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                                                  GenericTensorWrapper<double>(maxVal), m_maxBatchSize);
+                break;
             case THRESH_TRUNC:
-                Kernels::Host::trunc_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Host::trunc_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                                             m_maxBatchSize);
                 break;
             case THRESH_TOZERO:
-                Kernels::Host::tozero_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Host::tozero_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                                              m_maxBatchSize);
                 break;
             case THRESH_TOZERO_INV:
-                Kernels::Host::tozeroinv_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh), m_maxBatchSize);
+                Kernels::Host::tozeroinv_generic(inputWrapper, outputWrapper, GenericTensorWrapper<double>(thresh),
+                                                 m_maxBatchSize);
                 break;
         }
     }
 }
 
-void Threshold::operator()(hipStream_t stream, const Tensor &input, const Tensor &output,
-                           const Tensor &thresh, const Tensor &maxVal, eDeviceType device) {
+void Threshold::operator()(hipStream_t stream, const Tensor &input, const Tensor &output, const Tensor &thresh,
+                           const Tensor &maxVal, eDeviceType device) {
     // Verify that the tensors are located on the right device (CPU or GPU).
     CHECK_TENSOR_DEVICE(input, device);
     CHECK_TENSOR_DEVICE(output, device);
 
     // Ensure all tensors are using supported datatypes
-    CHECK_TENSOR_DATATYPES(input, eDataType::DATA_TYPE_U8, eDataType::DATA_TYPE_U16, eDataType::DATA_TYPE_S16, eDataType::DATA_TYPE_F32, eDataType::DATA_TYPE_F64);
-    CHECK_TENSOR_DATATYPES(output, eDataType::DATA_TYPE_U8, eDataType::DATA_TYPE_U16, eDataType::DATA_TYPE_S16, eDataType::DATA_TYPE_F32, eDataType::DATA_TYPE_F64);
+    CHECK_TENSOR_DATATYPES(input, eDataType::DATA_TYPE_U8, eDataType::DATA_TYPE_U16, eDataType::DATA_TYPE_S16,
+                           eDataType::DATA_TYPE_F32, eDataType::DATA_TYPE_F64);
+    CHECK_TENSOR_DATATYPES(output, eDataType::DATA_TYPE_U8, eDataType::DATA_TYPE_U16, eDataType::DATA_TYPE_S16,
+                           eDataType::DATA_TYPE_F32, eDataType::DATA_TYPE_F64);
     CHECK_TENSOR_DATATYPES(thresh, eDataType::DATA_TYPE_F64);
     CHECK_TENSOR_DATATYPES(maxVal, eDataType::DATA_TYPE_F64);
 
@@ -124,7 +131,8 @@ void Threshold::operator()(hipStream_t stream, const Tensor &input, const Tensor
     CHECK_TENSOR_COMPARISON(input.shape() == output.shape());
     CHECK_TENSOR_COMPARISON(input.dtype() == output.dtype());
     CHECK_TENSOR_COMPARISON(output.shape(output.layout().batch_index()) == input.shape(input.layout().batch_index()));
-    CHECK_TENSOR_COMPARISON(output.shape(output.layout().channels_index()) == input.shape(input.layout().channels_index()));
+    CHECK_TENSOR_COMPARISON(output.shape(output.layout().channels_index()) ==
+                            input.shape(input.layout().channels_index()));
     CHECK_TENSOR_COMPARISON(output.shape(output.layout().width_index()) == input.shape(input.layout().width_index()));
     CHECK_TENSOR_COMPARISON(output.shape(output.layout().height_index()) == input.shape(input.layout().height_index()));
     CHECK_TENSOR_COMPARISON(m_maxBatchSize == input.shape(input.layout().batch_index()));